requirements.txt
# Ana NLP Modelleri
torch
transformers
zemberek-python

# Dosya Okuyucular
PyMuPDF
python-docx

# Skorlama (Kosinüs)
scikit-learn
numpy